refactor(0.9.0): split SearchAgent god-class into 7 mixin modules
amx/search/agent.py was 3733 LOC with 70 methods spanning 6+
responsibilities. This commit splits the class into 7 mixin modules
under amx/search/_agent/ plus a thin facade.

* answering.py            (4 methods, 210 LOC) — synthesize_answer +
  provenance/confidence/actions
* deterministic.py        (6 methods, 549 LOC) — 6 _deterministic_*
  answer composers (no LLM)
* planning.py             (10 methods, 528 LOC) — plan/interpret/align/
  policy
* resolution.py           (12 methods, 597 LOC) — target + explicit
  mention resolution
* retrieval.py            (19 methods, 812 LOC) — _retrieve + live
  probes + row helpers
* session_memory.py       (10 methods, 201 LOC) — connection / memory
  / catalog state
* short_circuits.py       (6 methods, 338 LOC) — chitchat / meta /
  followup / tool-agent

Public API unchanged. SearchAgent now inherits from all 7 mixins;
cross-mixin self.* calls resolve via Python MRO. Method count
preserved (70 = 3 core + 67 mixin).

agent.py: 3733 -> 767 LOC (-79.5%).
