+++ README.md
@@ -25,6 +25,7 @@ Results from all agents are **merged** by an orchestrator using LLM reasoning, t
 - Write approved metadata back to the database as `COMMENT ON TABLE/VIEW/COLUMN` (write-back support)
 
 Recent release notes:
+- `v0.9.0`: **SearchAgent god-class split into mixin modules (S1 refactor).** `amx/search/agent.py` was 3733 LOC with 70 methods spanning 6+ responsibilities. v0.9.0 splits it into 7 mixin files under `amx/search/_agent/` (answering, deterministic, planning, resolution, retrieval, session_memory, short_circuits) plus a thin facade. agent.py is now 767 LOC (-79.5%). Public API unchanged — `SearchAgent.ask()` resolves all 70 methods transparently via Python MRO. Each cluster is independently testable; cross-mixin dependencies are documented in mixin docstrings.
 - `v0.8.7`: **Banner footer no longer duplicates the version.** v0.8.6's footer was `v0.8.6 • AI-inferred database descriptions`, but the version is also shown in the "AMX Interactive Session" info block right below the banner. Two adjacent panels stamping the same version was noise. Footer is now just the tagline; version stays in the session info block where it groups with Config / Database / LLM. Banner = identity, session info = runtime state.
 - `v0.8.6`: **Startup banner cleanup.** Dropped the redundant "Metadata Extraction System" subtitle (was saying the same thing as the title two lines above), swapped asterisks for box-drawing brackets so the framing matches the grid art instead of mixing vector glyphs, added a `v0.8.6  •  AI-inferred database descriptions` footer line auto-pulled from `amx.__version__`. Three-tier cyan hierarchy (tagline → art → footer) for readability.
 - `v0.8.5`: **`/run` Column scope — re-run AI inference on a single column.** User: "yalnızca tek 1 kolon için /run çalıştıramıyo muyuz". New 5th scope option (`Database / Schema / Asset / Column / Default`) drills schema → table → column and restricts the run to just that one column. Useful when one comment came out wrong (LLM picked the wrong meaning for `code`, say) and you want to regenerate without re-profiling the whole table. New `ScopeResult` dict subclass carries optional `column_overrides`; orchestrator.process_table consults it before missing-only / dedup filters. Equivalence pre-walk respects the override too. Existing scope shapes are unchanged plain dicts so other consumers keep working.
