+++ amx/__init__.py
@@ -1,6 +1,6 @@
 """AMX — Agentic Metadata Extractor."""
 
-__version__ = "0.8.7"
+__version__ = "0.9.0"
 
 __all__ = [
     "AMXApplication",
